+++ src/TestTrackingDiagrams/Tracking/MessageTrackerVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Tracking;
+
+public enum MessageTrackerVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}